+++ src/loginServices.test.js
@@ -71,7 +71,7 @@ jest.mock('./discoverServices', () => ({
 const mockStripesHubAPI = stripesHubAPI;
 jest.mock('localforage', () => ({
   getItem: jest.fn((str) => {
-    if (str === mockStripesHubAPI.HOST_LOCATION_KEY) {
+    if (str === mockStripesHubAPI.HOST_URL_KEY) {
       return Promise.resolve(null);
     }
     return Promise.resolve({ user: {} });
@@ -81,17 +81,10 @@ jest.mock('localforage', () => ({
 }));
 
 jest.mock('stripes-config', () => ({
-  config: {
-    tenantOptions: {
-      romulus: { name: 'romulus', clientId: 'romulus-application' },
-      remus: { name: 'remus', clientId: 'remus-application' },
-    }
-  },
-  okapi: {
-    authnUrl: 'https://authn.url',
-  },
   translations: { cs_CZ: 'cs-CZ', cs: 'cs-CZ', fr: 'fr', ar: 'ar', en_US: 'en-US', en_GB: 'en-GB' }
 }));
+
+const OKAPI = { authnUrl: 'https://authn.url', url: 'http://okapi-url' };
 
 // fetch success: resolve promise with ok == true and $data in json()
 const mockFetchSuccess = (data) => {
@@ -124,6 +117,12 @@ describe('createOkapiSession', () => {
         okapi: {
           currentPerms: [],
           url: 'okapiUrl'
+        },
+        config: {
+          tenantOptions: {
+            romulus: { name: 'romulus', clientId: 'romulus-application' },
+            remus: { name: 'remus', clientId: 'remus-application' },
+          }
         }
       }),
     };
@@ -206,13 +205,13 @@ describe('loadTranslations', () => {
     });
   });
 
-  describe('when localforage contains a hostLocation', () => {
-    it('fetches using the hostLocation from localforage', async () => {
-      const hostLocation = 'http://my-app-here';
+  describe('when localforage contains a hostUrl', () => {
+    it('fetches using the hostUrl from localforage', async () => {
+      const hostUrl = 'http://my-app-here';
       const locale = 'cs-CZ';
-      localforage.getItem.mockResolvedValueOnce(hostLocation);
+      localforage.getItem.mockResolvedValueOnce(hostUrl);
       await loadTranslations(store, locale, { cs: 'cs-CZ' });
-      expect(global.fetch).toHaveBeenCalledWith(`${hostLocation}/cs-CZ`);
+      expect(global.fetch).toHaveBeenCalledWith(`${hostUrl}/cs-CZ`);
     });
   });
 });
@@ -384,7 +383,7 @@ describe('validateUser', () => {
 
     mockFetchSuccess(data);
 
-    await validateUser('url', store, tenant, session);
+    await validateUser('url', store, tenant, session, null);
 
     const updatedSession = {
       user: { ...session.user, ...data.user },
@@ -575,7 +574,7 @@ describe('logout', () => {
       window.sessionStorage.setItem(IS_LOGGING_OUT, 'true');
 
       let res;
-      await logout('', store)
+      await logout('', store, null)
         .then(() => {
           res = true;
         });
@@ -593,12 +592,12 @@ describe('logout', () => {
       global.fetch = jest.fn().mockImplementation(() => Promise.resolve());
       const store = {
         dispatch: jest.fn(),
-        getState: jest.fn(),
+        getState: jest.fn().mockReturnValue({ config: { preserveConsole: false } }),
       };
       window.sessionStorage.clear();
 
       let res;
-      await logout('', store)
+      await logout('', store, null)
         .then(() => {
           res = true;
         });
@@ -615,12 +614,12 @@ describe('logout', () => {
       localStorage.setItem(SESSION_NAME, 'true');
       const store = {
         dispatch: jest.fn(),
-        getState: jest.fn(),
+        getState: jest.fn().mockReturnValue({ config: { preserveConsole: false } }),
       };
       window.sessionStorage.clear();
 
       let res;
-      await logout('', store)
+      await logout('', store, null)
         .then(() => {
           res = true;
         });
@@ -634,11 +633,12 @@ describe('logout', () => {
       localStorage.clear();
       const store = {
         dispatch: jest.fn(),
+        getState: jest.fn().mockReturnValue({ config: { preserveConsole: false } }),
       };
       window.sessionStorage.clear();
 
       let res;
-      await logout('', store)
+      await logout('', store, null)
         .then(() => {
           res = true;
         });
@@ -657,6 +657,7 @@ describe('logout', () => {
       global.fetch = jest.fn().mockImplementation(() => Promise.resolve());
       const store = {
         dispatch: jest.fn(),
+        getState: jest.fn().mockReturnValue({ okapi: { tenant: 'diku' }, config: { preserveConsole: false } }),
       };
       const rqc = {
         removeQueries: jest.fn(),
@@ -868,7 +869,7 @@ describe('unauthorizedPath functions', () => {
       };
 
       await requestUserWithPerms(
-        'http://okapi-url',
+        OKAPI,
         mockStore,
         'test-tenant',
         'token'
@@ -906,7 +907,7 @@ describe('unauthorizedPath functions', () => {
         })));
       fetchOverriddenUserWithPerms.mockResolvedValue(mockResponse);
 
-      await expect(requestUserWithPerms('okapiUrl', mockStore, 'tenant', true)).rejects.toEqual('Reject message');
+      await expect(requestUserWithPerms('okapiUrl', mockStore, 'tenant', 'token')).rejects.toEqual('Reject message');
       mockFetchCleanUp();
     });
   });